Java Client Connection Jar build changes
For the client connection jar provided to provide a way to call the
webservices from Java, the approach to building it has changed.

Gradle 'shadowJar' plugin is no longer used.
It is created using standard Gradle build jar and then Ant is used to
combine the Common code jar with the Client Connector jar into a single
jar for deployment